[Feat] 시간표 및 과목 구현

+++ src/main/java/allclear/crawl/ParsingGrade.java
@@ -20,7 +20,7 @@ public class ParsingGrade {
         String tmpNum = "64"; // 총 이수학점 임시 점수
         semesterAverageGrade = "";
         grade = new Grade();
-        grade.setTotalCredit(Long.parseLong(tmpNum));
+        grade.setTotalCredit(Integer.parseInt(tmpNum));
         grade.setAverageGrade(averageGrade);
         ArrayList<SemesterSubject> tmpList = new ArrayList<>();
